config/requirements.txt
Django==2.2.19
django_filter==21.1
djangorestframework==3.13.1
requests==2.27.1
